Minor README adjustments
The TOC was missing an entry.  While fixing it, I discovered a [vim
plugin that can do it for me][plugin].

Also, multiple TOC in a single document is probably a bad idea, so
remove classes and defined defined types sub-TOC…  All the Reference
section will be moved in an auto-generated REFERENCE file at some point.

[plugin]: https://github.com/mzlogin/vim-markdown-toc